+++ aasrp/app_settings.py
@@ -1,5 +1,5 @@
 """
-our app setting
+App settings
 """
 
 # Django
@@ -8,37 +8,40 @@ from django.apps import apps
 # AA SRP
 from aasrp.utils import clean_setting
 
-# AA-GDPR
-AVOID_CDN = clean_setting("AVOID_CDN", False)
-
 # SRP Team Discord Channel
 AASRP_SRP_TEAM_DISCORD_CHANNEL = clean_setting(
     name="AASRP_SRP_TEAM_DISCORD_CHANNEL", default_value=None, required_type=int
 )
 
 
-def avoid_cdn() -> bool:
+def allianceauth_discordbot_installed() -> bool:
     """
-    check if we should aviod CDN usage
-    :return: bool
-    """
-
-    return AVOID_CDN
-
-
-def allianceauth_discordbot_active():
-    """
-    check if allianceauth-dicordbot is installed and active
+    Check if allianceauth-discordbot is installed and active
     :return:
     """
 
     return apps.is_installed("aadiscordbot")
 
 
-def aa_discordnotify_active():
+def aa_discordnotify_installed() -> bool:
     """
-    check if allianceauth-dicordbot is installed and active
+    Check if allianceauth-discordbot is installed and active
     :return:
     """
 
     return apps.is_installed("discordnotify")
+
+
+def discordproxy_installed() -> bool:
+    """
+    Check if discordproxy is installed by trying to import the DiscordClient
+    :return:
+    """
+
+    try:
+        # Third Party
+        from discordproxy.client import DiscordClient  # noqa: F401
+    except ModuleNotFoundError:
+        return False
+    else:
+        return True