+++ aasrp/auth_hooks.py
@@ -20,7 +20,7 @@ class AaSrpMenuItem(MenuItemHook):  # pylint: disable=too-few-public-methods
     """
 
     def __init__(self):
-        # setup menu entry for sidebar
+        # Setup menu entry for sidebar
         MenuItemHook.__init__(
             self,
             _(__title__),
@@ -31,7 +31,7 @@ class AaSrpMenuItem(MenuItemHook):  # pylint: disable=too-few-public-methods
 
     def render(self, request):
         """
-        check if the user has the permission to view this app
+        Check if the user has the permission to view this app
         :param request:
         :return:
         """
@@ -48,7 +48,7 @@ class AaSrpMenuItem(MenuItemHook):  # pylint: disable=too-few-public-methods
 @hooks.register("menu_item_hook")
 def register_menu():
     """
-    register our menu item
+    Register our menu item
     :return:
     """
 
@@ -58,7 +58,7 @@ def register_menu():
 @hooks.register("url_hook")
 def register_urls():
     """
-    register our basu url
+    Register our basu url
     :return:
     """